requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.12
numpy==2.4.4
pandas==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
six==1.17.0
urllib3==2.6.3
